requirements.txt
aniso8601==9.0.1
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.3
Deprecated==1.2.13
Flask==2.0.2
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
Flask-Restless==0.17.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.26.1
mimerender==0.6.0
pycparser==2.21
PyGithub==1.55
PyJWT==2.3.0
PyNaCl==1.4.0
python-dateutil==2.8.2
python-mimeparse==1.6.0
pytz==2021.3
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.27
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
